requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
appdirs==1.4.4
asgiref==3.11.1
asyncssh==2.22.0
atpublic==7.0.0
attrs==25.4.0
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
cryptography==43.0.3
curl_cffi==0.13.0
cycler==0.12.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
dpath==2.2.0
dulwich==0.25.2
dvc==3.66.1
dvc-data==3.18.2
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.2.0
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
email-validator==2.3.0
entrypoints==0.4
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.9.0
fastar==0.8.0
filelock==3.20.3
flatten-dict==0.4.2
flufl.lock==9.0.0
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.1.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.61.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grandalf==0.8
grpcio==1.76.0
grpcio-status==1.71.2
gto==1.9.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httptools==0.7.1
httpx==0.28.1
hydra-core==1.3.2
idna==3.11
importlib_metadata==8.7.1
iterative-telemetry==0.0.10
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
kiwisolver==1.4.9
kombu==5.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
multidict==6.7.0
multitasking==0.0.12
networkx==3.6.1
numpy==2.4.1
oauth2client==4.1.3
omegaconf==2.3.0
openai==2.16.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
orjson==3.11.5
packaging==25.0
pandas==3.0.0
pathspec==0.12.1
peewee==3.19.0
pillow==12.1.0
platformdirs==4.5.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydot==4.0.1
PyDrive2==1.21.3
pygit2==1.19.1
Pygments==2.19.2
pygtrie==2.5.0
pyOpenSSL==24.2.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
ruamel.yaml==0.19.1
scikit-learn==1.8.0
scipy==1.17.0
scmrepo==3.6.1
seaborn==0.13.2
semver==3.0.4
sentry-sdk==2.49.0
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
sqltrie==0.11.2
starlette==0.50.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tomlkit==0.14.0
tqdm==4.67.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
vine==5.1.0
voluptuous==0.16.0
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wheel==0.45.1
wrapt==1.17.3
yarl==1.22.0
yfinance==1.1.0
zc.lockfile==4.0
zipp==3.23.0
